Deleting matplotlib based manhattan.py. Replacing it with tk based version
